requirements.txt
azure-functions>=1.15.0
